+++ internal/linter/testdata/simple/clean.go
@@ -1,0 +1,7 @@
+package simple
+
+import "github.com/lex00/wetwire-aws-go/resources/s3"
+
+var MyBucket = s3.Bucket{
+	BucketName: "my-bucket",
+}